pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = ["MarkCesium <cesiummark01@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2.*"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^2.3"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
coverage = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
